validate application id
