+++ pkg/connectorrp/renderers/daprstatestores/renderer.go
@@ -42,6 +42,10 @@ func (r *Renderer) Render(ctx context.Context, dm conv.DataModelInterface, optio
 
 	properties := resource.Properties
 
+	err := renderers.ValidateApplicationID(properties.Application)
+	if err != nil {
+		return renderers.RendererOutput{}, err
+	}
 	if r.StateStores == nil {
 		return renderers.RendererOutput{}, errors.New("must support either kubernetes or ARM")
 	}
