+++ pkg/connectorrp/renderers/rabbitmqmessagequeues/renderer.go
@@ -28,6 +28,10 @@ func (r Renderer) Render(ctx context.Context, dm conv.DataModelInterface, option
 
 	properties := resource.Properties
 
+	err := renderers.ValidateApplicationID(properties.Application)
+	if err != nil {
+		return renderers.RendererOutput{}, err
+	}
 	if properties.Secrets == (datamodel.RabbitMQSecrets{}) || properties.Secrets.ConnectionString == "" {
 		return renderers.RendererOutput{}, renderers.NewClientErrInvalidRequest("secrets must be specified")
 	}
