+++ pkg/connectorrp/renderers/rediscaches/renderer.go
@@ -34,6 +34,11 @@ func (r Renderer) Render(ctx context.Context, dm conv.DataModelInterface, option
 	properties := resource.Properties
 	secretValues := getProvidedSecretValues(properties)
 
+	err := renderers.ValidateApplicationID(properties.Application)
+	if err != nil {
+		return renderers.RendererOutput{}, err
+	}
+
 	if resource.Properties.Resource == "" {
 		return renderers.RendererOutput{
 			Resources: []outputresource.OutputResource{},
